INT changed to BIGINT

+++ migrations/1_init.up.sql
@@ -1,6 +1,6 @@
 CREATE TABLE IF NOT EXISTS users
 (
-    id        INT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
+    id        BIGINT GENERATED ALWAYS AS IDENTITY PRIMARY KEY,
     email     TEXT NOT NULL UNIQUE,
     pass_hash bytea NOT NULL
 );
@@ -8,7 +8,7 @@ CREATE INDEX IF NOT EXISTS idx_email ON users (email);
 
 CREATE TABLE IF NOT EXISTS apps
 (
-    id     INT PRIMARY KEY,
+    id     BIGINT PRIMARY KEY,
     name   TEXT NOT NULL UNIQUE,
     secret TEXT NOT NULL UNIQUE
 );
